Time: 8 ms (69.54%), Space: 8.7 MB (36.60%) - LeetHub

+++ 0013-roman-to-integer/0013-roman-to-integer.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int romanToInt(string s) {
-        unordered_map<char,int>map;
+        unordered_map<char,short>map;
         map['I'] = 1;
         map['V'] = 5;
         map['X'] = 10;
@@ -10,7 +10,7 @@ public:
         map['D'] = 500;
         map['M'] = 1000;
 
-        int ans = 0;
+        short ans = 0;
 
         for (short i = 0; i < s.size();i++){
             if(map[s[i]]<map[s[i+1]]){
